// src/app/interfaces/Iaccion.ts
export interface Iaccion {
    simbolo: string;
    ultimoOperado: number;
    variacionDiaria: number;
    cantidadCompra: number;
    precioCompra: number;
    precioVenta: number;
    cantidadVenta: number;
    apertura: number;
    minimo: number;
    maximo: number;
    ultimoCierre: number;
    montoOperado: number;
}